refactored to have PageLayout pass rows[] to InfoTable. Api calls made in parent component, PageLayout
